apps/ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
paddlepaddle==2.6.1
paddleocr==2.7.3
google-generativeai==0.5.4
python-multipart==0.0.9
pydantic==2.7.4
python-dotenv==1.0.1
